requirements.txt
flet
fleting
supabase
python-dotenv
